Create crud-bateas lib
