+++ truck-fleet-manager/angular.json
@@ -200,6 +200,39 @@
           }
         }
       }
+    },
+    "crud-bateas": {
+      "projectType": "library",
+      "root": "projects/crud-bateas",
+      "sourceRoot": "projects/crud-bateas/src",
+      "prefix": "lib",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-angular:ng-packagr",
+          "options": {
+            "project": "projects/crud-bateas/ng-package.json"
+          },
+          "configurations": {
+            "production": {
+              "tsConfig": "projects/crud-bateas/tsconfig.lib.prod.json"
+            },
+            "development": {
+              "tsConfig": "projects/crud-bateas/tsconfig.lib.json"
+            }
+          },
+          "defaultConfiguration": "production"
+        },
+        "test": {
+          "builder": "@angular-devkit/build-angular:karma",
+          "options": {
+            "tsConfig": "projects/crud-bateas/tsconfig.spec.json",
+            "polyfills": [
+              "zone.js",
+              "zone.js/testing"
+            ]
+          }
+        }
+      }
     }
   }
 }
