+++ truck-fleet-manager/projects/crud-bateas/README.md
@@ -1,0 +1,25 @@
+# CRUDBateas
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 16.2.0.
+
+## Code scaffolding
+
+Run `ng generate component component-name --project crud-bateas` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project crud-bateas`.
+
+> Note: Don't forget to add `--project crud-bateas` or else it will be added to the default project in your `angular.json` file.
+
+## Build
+
+Run `ng build crud-bateas` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Publishing
+
+After building your library with `ng build crud-bateas`, go to the dist folder `cd dist/crud-bateas` and run `npm publish`.
+
+## Running unit tests
+
+Run `ng test crud-bateas` to execute the unit tests via [Karma](https://karma-runner.github.io).
+
+## Further help
+
+To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
