+++ truck-fleet-manager/projects/crud-bateas/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/crud-bateas",
+  "lib": {
+    "entryFile": "src/index.ts"
+  }
+}
