+++ truck-fleet-manager/projects/crud-bateas/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "crud-bateas",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^16.2.0",
+    "@angular/core": "^16.2.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "sideEffects": false
+}
