+++ truck-fleet-manager/projects/crud-bateas/src/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of crud-bateas
+ */
+
+export * from './lib/crud.bateas.module';
