+++ truck-fleet-manager/projects/crud-bateas/src/lib/components/batea-list-component/batea.list.component.html
@@ -1,7 +1,13 @@
 <h1>Gestión de Bateas</h1>
 
-<fm-form [entityName]="'Batea'" [fields]="['patente']" [editMode]="editMode" [title]="formTitle"
-  (formSubmit)="onSubmit($event)" (formEdit)="setbateaForm($event)"></fm-form>
+<fm-form
+  [entityName]="'Batea'"
+  [fields]="['patente']"
+  [editMode]="editMode"
+  [title]="formTitle"
+  (formSubmit)="onSubmit($event)"
+  (formEdit)="setbateaForm($event)"
+></fm-form>
 
 <fm-table
   [data]="bateasList$"
