+++ truck-fleet-manager/projects/crud-bateas/src/lib/components/batea-list-component/batea.list.component.scss
