+++ truck-fleet-manager/projects/crud-bateas/src/lib/components/index.ts
@@ -1,0 +1,1 @@
+export * from './batea-list-component/batea.list.component';
