+++ truck-fleet-manager/projects/crud-bateas/src/lib/crud.bateas.module.ts
@@ -1,24 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
-
-import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 
-import { CommonUiModule } from '../../../projects/common-ui/src';
-import { BateaRoutingModule } from './batea-routing.module';
-import { BateaComponent } from './components/batea-component/batea.component';
+import { CommonUiModule } from 'projects/common-ui/src';
+
+import { CrudBateasRoutingModule } from './crud.bateas.routing.module';
+import { BateaListComponent } from './components/batea-list-component/batea.list.component';
 
 @NgModule({
   declarations: [
-    BateaComponent
+    BateaListComponent
   ],
   imports: [
     CommonModule,
-    BateaRoutingModule,
+    CrudBateasRoutingModule,
     HttpClientModule,
     MatSnackBarModule,
     CommonUiModule
   ]
 })
-export class BateaModule { }
+export class CrudBateasModule { }
