+++ truck-fleet-manager/projects/crud-bateas/src/lib/models/batea.model.ts
