+++ truck-fleet-manager/projects/crud-bateas/src/lib/models/index.ts
